pyproject.toml
[project]
name = "geronimo"
version = "0.1.0"
description = "MLOps deployment platform - automate ML model deployments to AWS with Terraform, Docker, and CI/CD"
readme = "README.md"
authors = [
    { name = "Chris Sweet", email = "chris.sweet0320@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "jinja2>=3.1.6",
    "pandas>=2.0.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "typer>=0.21.1",
]

[project.optional-dependencies]
mlflow = [
    "mlflow>=2.10.0",
]
databases = [
    "snowflake-connector-python>=3.0.0",
    "psycopg2-binary>=2.9.0",
    "pyodbc>=5.0.0",
]
all = [
    "geronimo[mlflow]",
    "geronimo[databases]",
]

[project.scripts]
geronimo = "geronimo.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "pytest-cov>=7.0.0",
]

